+++ src/main/java/blackjack/domain/Player.java
@@ -1,34 +1,15 @@
 package blackjack.domain;
 
-import java.util.List;
 
-public class Player {
+public class Player extends Participant {
     private final String name;
-
-    private final Cards drawnCards;
 
     public Player(String name) {
         this.name = name;
-        this.drawnCards = new Cards();
     }
 
     public String getName() {
         return name;
     }
 
-    public void receiveOneCard(Card card) {
-        drawnCards.addCard(card);
-    }
-
-    public List<String> getCardNames() {
-        return drawnCards.getCardNames();
-    }
-
-    public boolean isBust() {
-        return drawnCards.sumScore() > 21;
-    }
-
-    public int calculateTotalScore() {
-        return drawnCards.sumScore();
-    }
 }
